Add @media print specific styling to fix problem
The way of styling the page is different than before. Before, the user would see all the changed done to the page to make it printable to a4 format. Now, however, everything is done in the @media print css block. This way improves user experience and it also 'fixes' the problem of getting stuck on a weirdly formatted page if the user presses cancel instead of print