Refactor initGenerate function -> rewrite fetchInvoiceByID func in service
The part of the generateInvoice function that did stuff with html and nunjucks has been moved to the htmlService class, while the part that fetched all necessary objects has been moved to the fetchInvoiceByID function. Also, not only has the code been moved, there is also a change in the code functionality itself. Right now, fetchInvoiceByID returns an invoiceDTO, containing all the information neccessary to create and generate an invoice