Rewrite money calculations using ezMoney library, add extra fields to jobDTO regarding costs
The ezmoney library has been used to calculate how much money is owed, instead of just using the number datatype. The number datatype is not suitable for money i believe. Extra fields have been added to different dto's (job, cameraman, etc.) regarding the costs. This was done because before, the whole job object was sent which made it possible to call the calculate() methods in the template. Now, however DTO's are passed to the temlpates, which means the previously used methods weren't available. So they had to be added as attributes. Note: in the templates, the toFixed(2) method is used on the totalCosts, vatCosts, etc. fields to ensure that they always display two decimals. This was done because if a cost (e.g. cameraman cost) = 100.10, the trailing zero's (0) got removed.